requirements.txt
requests==2.31.0
bs4==0.0.1
openai==0.27.7
tiktoken==0.4.0
python-dotenv==1.0.0
langchain==0.0.177
chromadb==0.3.23